Reorganize API namespaces - first cut
