+++ src/testcentric.engine.api/ITestEngine.cs
@@ -4,6 +4,7 @@
 // ***********************************************************************
 
 using System;
+using TestCentric.Engine.Services;
 
 namespace TestCentric.Engine
 {
