+++ src/testcentric.engine.api/ITestEngineRunner.cs
@@ -4,7 +4,6 @@
 // ***********************************************************************
 
 using System;
-using TestCentric.Engine.Extensibility;
 
 namespace TestCentric.Engine
 {
