+++ src/testcentric.engine.api/Services/IRuntimeFrameworkService.cs
@@ -3,7 +3,7 @@
 // Licensed under the MIT License. See LICENSE file in root directory.
 // ***********************************************************************
 
-namespace TestCentric.Engine
+namespace TestCentric.Engine.Services
 {
     /// <summary>
     /// Implemented by a type that provides information about the
