+++ src/testcentric.engine.api/Services/ITestAgency.cs
@@ -3,7 +3,7 @@
 // Licensed under the MIT License. See LICENSE file in root directory.
 // ***********************************************************************
 
-namespace TestCentric.Engine
+namespace TestCentric.Engine.Services
 {
     /// <summary>
     /// The ITestAgency interface is implemented by a TestAgency in
