+++ src/testcentric.engine.api/Services/ITestFilterBuilder.cs
@@ -3,7 +3,7 @@
 // Licensed under the MIT License. See LICENSE file in root directory.
 // ***********************************************************************
 
-namespace TestCentric.Engine
+namespace TestCentric.Engine.Services
 {
     /// <summary>
     /// Interface to a TestFilterBuilder, which is used to create TestFilters
